Remove categories and subcategories queries
This commit removes usage of Categories queries in order to support
postgres. The Categories in postgres is an enum instead of a table and
we don't need to use the table in sqlite because the value are hardcoded
in the html and are unlikely to ever change.

This commit breaks the thumbnails in the index. Other functionality
needs to be tested before merging.