Fix filelist folder show/hide logic
It only toggled the "hidden" attributes of the childs, however if
you closed a subfolder, then closed the parent folder, all childs of the
parent would be toggled, leaving the subfolder that we closed open.

Now the state is stored on the data-filelist-open attribute, and the
hidden state is set based on it.