Apply max-width/height to all description images
Max-width & max-height only applied to images inside another element in the torrent description. It's fine because description images seem to always be in a \<p\>\<p/\>, but what if they weren't?
The max-width & max-height now apply to all images in the torrent description (.torrent-info-box img) instead of one in another element (.torrent-info-box > * > img), that way no image may ever go outside the bounds! Not on it's own at least.